requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
bootstrap-admin==0.4.4
cffi==1.15.0
coreapi==2.0.9
coverage==4.3.4
cryptography==35.0.0
Django==3.2.8
django-admin-rangefilter==0.8.1
django-audit-log==0.7.0
django-cors-headers==3.10.0
django-discover-runner==1.0
django-filter==21.1
django-rest-auth==0.9.5
django-rest-swagger==2.1.0
django-role-permissions==3.1.1
djangorestframework==3.12.4
drf-extra-fields==3.1.1
drf-spectacular==0.20.2
Fabric3==1.13.1.post1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.2
inflection==0.5.1
iso8601==0.1.11
itypes==1.1.0
jsonschema==4.1.2
numpy==1.21.3
openapi-codec==1.1.7
paramiko==2.1.2
Pillow==8.4.0
promise==2.3
psycopg2==2.9.1
pyasn1==0.2.2
pycparser==2.17
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
requests==2.12.1
Rx==1.6.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing==3.5.3.0
uritemplate==3.0.0
